getFov and projectColorToDepth added to android API

+++ wrappers/android/librealsense/src/main/java/com/intel/realsense/librealsense/Pixel.java
@@ -1,15 +1,15 @@
 package com.intel.realsense.librealsense;
 
 public class Pixel {
-    public int mX;
-    public int mY;
+    public float mX;
+    public float mY;
 
     public Pixel(){
-        mX = 0;
-        mY = 0;
+        mX = 0.f;
+        mY = 0.f;
     }
 
-    public Pixel(int x, int y){
+    public Pixel(float x, float y){
         mX = x;
         mY = y;
     }
